Switch to Streams (to revert)
I think that to the best way to deal with different encodings is to avoid
dealing with them, assuming we're using an ASCII extension (which we
usually are). But we need to do that when dealing with the input!

And InputStreams don't give strings back... actually,
DataInputStream.readLine does. But a better approach is to use
a pure 8bit encoding, like latin1, which will misunderstand
extended characters, but never produce an error. Moreover, it will
pass misunderstood characters through just fine, I think.

So this attempt is to revert.